requirements.txt
python-chess==1.999
stockfish==3.28.0
ollama==0.2.1
fastapi==0.111.1
uvicorn[standard]==0.30.1
pytest==8.4.1
pytest-mock==3.14.1
python-multipart==0.0.9
